moved viewmodel to its own file
